renaming of Subscription files to Publisher files
